Core Automated Testing (#72)
Added automated testing for the core functions common to every filters.

Tested under:

    FreeBSD 12.0
        Redis 4.0.14
        Boost 1.71.0
        clang++ 8.0.1
        CMake 3.15.2
        Python 3.6.9
    Ubuntu 19.04
        Redis 5.0.3
        Boost 1.71.0
        g++ 8.3.0
        CMake 3.13.4
        Python 3.7.3
        Valgrind 3.14.0

Test list:

    check_start_stop
    check_start_wrong_conf
    check_pid_file
    check_socket_create_delete
    check_socket_connection
    check_socket_monitor_create_delete
    check_socket_monitor_connection
    check_start_no_conf

I put core testing before manager and filters because a working core is mandatory to test both.

Other modifications:

    FIX: Check config is a JSON object before accessing members in filters' generator.
    FIX: Missing return False in manager testing.
    FIX: Error prints in manager testing utils. Used logger.
    FIX: Memory error allocating using offsetof in darwin::Session. Now calculating using sizeof.
    FIX: Missing free in fconnection before writing new entry to REDIS.
    FIX: Missing free in RedisManager in case of connection to a READONLY slave.
    FIX: Logged valgrind test output was stdout, valgrind write on stderr.
    Changed valgrind tests verbosity to quiet to ensure only errors are logged.
    Moved filter test between Core and Manager because working filters are necessary for manager testing.
    Added more details in pid file check logs.

+++ tests/tools/filter.py
@@ -2,10 +2,10 @@ import subprocess
 import os
 import logging
 from time import sleep
-from conf import DEFAULT_FILTER_PATH
+from conf import DEFAULT_FILTER_PATH, VALGRIND_MEMCHECK
 
 
-class Filter:
+class Filter():
 
     def __init__(self, path=None, config_file=None, filter_name="filter", socket_path=None, monitoring_socket_path=None, pid_file=None, output="NONE", next_filter_socket_path="no", nb_thread=1, cache_size=0, thresold=101, log_level="-z"):
         self.filter_name = filter_name
@@ -16,11 +16,14 @@ class Filter:
         self.pid = pid_file if pid_file else "/tmp/{}.pid".format(filter_name)
         self.cmd = [self.path, self.filter_name, self.socket, self.config, self.monitor, self.pid, output, next_filter_socket_path, str(nb_thread), str(cache_size), str(thresold), log_level]
         self.process = None
-        self.error_code = 99 # For valgrind testing
+        self.error_code = 99 # For valgrind testing
 
     def __del__(self):
         if self.process and self.process.poll() is None:
-            self.stop()
+            if VALGRIND_MEMCHECK is False:
+                self.stop()
+            else:
+                self.valgrind_stop()
         self.clean_files()
 
     def start(self):
@@ -39,10 +42,13 @@ class Filter:
         return True
 
     def valgrind_start(self):
+        if VALGRIND_MEMCHECK is False:
+            self.start()
+            return
         command = ['valgrind',
                    '--tool=memcheck',
                    '--leak-check=yes',
-                   '-v', # Quiet mode so it doesn't pollute the output
+                   '-q', # Quiet mode so it doesn't pollute the output
                    '--error-exitcode={}'.format(self.error_code), #If valgrind report error on the run, return this exitcode.
                    ] + self.cmd
 
@@ -50,10 +56,13 @@ class Filter:
         sleep(3)
 
     def valgrind_stop(self):
+        if VALGRIND_MEMCHECK is False:
+            self.stop()
+            return True
         sleep(3)
         self.process.terminate()
         try:
-            self.process.wait(15) # Valgrind can take some times, so it need a generous timeout
+            self.process.wait(15) # Valgrind can take some times, so it need a generous timeout
         except subprocess.TimeoutExpired:
             logging.error("Unable to stop filter properly... Killing it.")
             self.process.kill()
@@ -64,7 +73,7 @@ class Filter:
         if self.process.returncode == self.error_code :
             out, err = self.process.communicate()
             logging.error("Valgrind returned error code: {}".format(self.process.returncode))
-            logging.error("Valgrind error: {}".format(out))
+            logging.error("Valgrind error: {}".format(err))
             return False
 
         return True
